feat: move track row to component
