+++ src/components/TrackRow.vue
@@ -1,0 +1,69 @@
+<template>
+  <div class="title">
+    <!-- <a class="disabled" ng-if="song.disabled" ng-click="copyrightNotice()"> song.title </a> -->
+    <vue-feather v-if="!isRedHeart(song.id)" type="heart" size="18" stroke-width="1" stroke="#666666" @click="setRedHeart(toRaw(song), true)" />
+    <vue-feather v-if="isRedHeart(song.id)" type="heart" fill="red" stroke="red" size="18" @click="setRedHeart(toRaw(song), false)" />
+
+    <a @click="play(song)">{{ song.title }}</a>
+  </div>
+  <div class="artist">
+    <a @click="$router.push(`/playlist/${props.song.artist_id}`)">{{ props.song.artist }}</a>
+  </div>
+  <div class="album">
+    <a @click="$router.push(`/playlist/${props.song.album_id}`)">{{ props.song.album }}</a>
+  </div>
+  <div class="tools">
+    <a v-show="song.options" :title="t('_ADD_TO_QUEUE')" class="detail-add-button" @click="addToPlay(song)"><span class="icon li-add" /></a>
+    <a v-show="song.options" :title="t('_ADD_TO_PLAYLIST')" class="detail-fav-button" @click="showModal('AddToPlaylist', { tracks: [song] })">
+      <span class="icon li-songlist" />
+    </a>
+    <a
+      v-show="song.options && (is_mine == '1' || is_local)"
+      :title="t('_REMOVE_FROM_PLAYLIST')"
+      class="detail-delete-button"
+      @click="removeSongFromPlaylist(song.id, list_id)">
+      <span class="icon li-del" />
+    </a>
+    <a v-show="song.options && !is_local" :title="t('_ORIGIN_LINK')" class="source-button" @click="openUrl(song.source_url)">
+      <span class="icon li-link" />
+    </a>
+  </div>
+</template>
+<script setup lang="ts">
+const props = defineProps<{
+  song: any;
+  is_mine?: string;
+  is_local?: string;
+  list_id?: string;
+}>();
+import { inject, toRaw } from 'vue';
+import { useI18n } from 'vue-i18n';
+import notyf from '../services/notyf';
+import useRedHeart from '../composition/redheart';
+
+import { l1Player } from '../services/l1_player';
+const { isRedHeart, setRedHeart, removeTrackFromMyPlaylistByUpdateRedHeart } = useRedHeart();
+
+const showModal = inject('showModal');
+
+const { t } = useI18n();
+
+const play = (song: any) => {
+  l1Player.addTrack(toRaw(song));
+  l1Player.playById(song.id);
+};
+const openUrl = (url: string) => {
+  window.open(url, '_blank')?.focus();
+};
+const addToPlay = (song: any) => {
+  l1Player.addTrack(toRaw(song));
+  notyf.success(t('_ADD_TO_QUEUE_SUCCESS'));
+};
+const removeSongFromPlaylist = async (track_id: string, list_id?: string) => {
+  if (!list_id) {
+    return;
+  }
+  await removeTrackFromMyPlaylistByUpdateRedHeart(track_id, list_id);
+  notyf.success(t('_REMOVE_SONG_FROM_PLAYLIST_SUCCESS'));
+};
+</script>
