+++ src/views/Playlist.vue
@@ -106,35 +106,7 @@
           @drop="onPlaylistSongDrop(list_id, song, $event)"
           @mouseenter="song.options = true"
           @mouseleave="song.options = undefined">
-          <div class="title">
-            <!-- <a class="disabled" ng-if="song.disabled" ng-click="copyrightNotice()"> song.title </a> -->
-            <vue-feather v-if="!isRedHeart(song.id)" type="heart" size="18" stroke-width="1" stroke="#666666" @click="setRedHeart(toRaw(song), true)" />
-            <vue-feather v-if="isRedHeart(song.id)" type="heart" fill="red" stroke="red" size="18" @click="setRedHeart(toRaw(song), false)" />
-
-            <a add-and-play="song" @click="play(song)">{{ song.title }}</a>
-          </div>
-          <div class="artist">
-            <a @click="showPlaylist(song.artist_id)">{{ song.artist }}</a>
-          </div>
-          <div class="album">
-            <a @click="showPlaylist(song.album_id)">{{ song.album }}</a>
-          </div>
-          <div class="tools">
-            <a v-show="song.options" :title="t('_ADD_TO_QUEUE')" class="detail-add-button" @click="addToPlay(song)"><span class="icon li-add" /></a>
-            <a v-show="song.options" :title="t('_ADD_TO_PLAYLIST')" class="detail-fav-button" @click="showModal('AddToPlaylist', { tracks: [song] })">
-              <span class="icon li-songlist" />
-            </a>
-            <a
-              v-show="song.options && (is_mine == '1' || is_local)"
-              :title="t('_REMOVE_FROM_PLAYLIST')"
-              class="detail-delete-button"
-              @click="removeSongFromPlaylist(song.id, list_id)">
-              <span class="icon li-del" />
-            </a>
-            <a v-show="song.options && !is_local" :title="t('_ORIGIN_LINK')" class="source-button" @click="openUrl(song.source_url)">
-              <span class="icon li-link" />
-            </a>
-          </div>
+          <TrackRow :song="song" :is_local="is_local" :is_mine="is_mine" :list_id="list_id"></TrackRow>
         </DragDropZone>
       </ul>
     </div>
@@ -153,6 +125,7 @@ import notyf from '../services/notyf';
 import $event from '../services/EventService';
 import DragDropZone from '../components/DragDropZone.vue';
 import useRedHeart from '../composition/redheart';
+import TrackRow from '../components/TrackRow.vue';
 
 const { t } = useI18n();
 const router = useRouter();
@@ -231,10 +204,7 @@ const saveAsMyPlaylist = async (list_id) => {
   await MediaService.clonePlaylist(list_id, 'my');
   notyf.success(t('_ADD_TO_PLAYLIST_SUCCESS'));
 };
-const removeSongFromPlaylist = async (track_id, list_id) => {
-  await removeTrackFromMyPlaylistByUpdateRedHeart(track_id, list_id);
-  notyf.success(t('_REMOVE_SONG_FROM_PLAYLIST_SUCCESS'));
-};
+
 const addMylist = async () => {
   await l1Player.addTracks(toRaw(songs));
   notyf.success(t('_ADD_TO_QUEUE_SUCCESS'));
